+++ utils/GEOnex_Rover_ESP32/include/mqtt_manager.h
@@ -6,7 +6,7 @@
 void connectMQTT();
 void publishGPSData(float latitude, float longitude, int satellites, String time);
 void publish_wifi_strength();
-void publishData(String deviceId, String status, float latitude, float longitude, int satellites,
+void publishData(String deviceId, String status, double latitude, double longitude, int satellites,
                  String time, float colat, float colon, int battery, int wifi);
 
 void publishStatus(String status, int battery, int wifi);
